backend/question table/deleting question/requirements.txt
google-cloud-firestore==2.3.2
flask==2.0.1

